+++ example/src/views/characters/index.js
@@ -49,6 +49,7 @@ export default new AbstractCollectionBuilder({
   .genDetail({
     notes: true,
     picture: true,
+    files: true,
   })
   .genListSettings({
     defaultSortPreference: 'firstName',