+++ packages/q3-admin/src/components/Header/Header.test.jsx
@@ -1,31 +1,22 @@
 import React from 'react';
-import Grid from '@material-ui/core/Grid';
 import Header from './Header';
 
-const getGridItems = (el, expectedNumberOfResults) =>
-  expect(el.find(Grid)).toHaveLength(
-    expectedNumberOfResults,
-  );
-
-const getTitleSpacing = (el, expectedSpaceValue) =>
-  expect(el.find('#q3-app-title').props()).toHaveProperty(
-    'spacing',
-    expectedSpaceValue,
-  );
+const getSubNavContainer = (props, expectLength) =>
+  expect(
+    global
+      .shallow(<Header title="test" {...props} />)
+      .find('#q3-app-subnav'),
+  ).toHaveLength(expectLength);
 
 describe('Admin>Header', () => {
-  it('should not render backComponent', () => {
-    const el = global.shallow(<Header title="test" />);
-    getGridItems(el, 5);
-    getTitleSpacing(el, 0);
-  });
+  it('should not render navComponent', () =>
+    getSubNavContainer({}, 0));
 
-  it('should render backComponent', () => {
-    const el = global.shallow(
-      <Header title="test" backComponent={<div />} />,
-    );
-
-    getGridItems(el, 6);
-    getTitleSpacing(el, 1);
-  });
+  it('should render navComponent', () =>
+    getSubNavContainer(
+      {
+        navComponent: <div />,
+      },
+      1,
+    ));
 });