+++ packages/q3-admin/src/containers/DetailHeader/DetailHeader.jsx
@@ -1,8 +1,41 @@
 import React from 'react';
+import { Avatar } from 'q3-ui-filemanager';
 import Back from '../back';
 import Header from '../../components/Header';
+import { useAppContext } from '../../hooks';
 import { useTitle } from '../../hooks';
-import { Store } from '../state';
+import { Dispatcher, Store } from '../state';
+import FeaturedPhoto from '../FeaturedPhoto';
+
+export const DetailHeaderBackComponent = (props) => {
+  const { data } = React.useContext(Store);
+  const { patch } = React.useContext(Dispatcher);
+  const { can } = useAppContext(props);
+  const fn = patch();
+
+  return (
+    <>
+      <Back />
+      {can('picture') ? (
+        <>
+          <FeaturedPhoto
+            component={Avatar}
+            src={data.photo}
+            update={fn}
+          />
+        </>
+      ) : (
+        // offset missing avatar
+        <div
+          style={{
+            display: 'inline-block',
+            height: 50,
+          }}
+        />
+      )}
+    </>
+  );
+};
 
 const DetailHeader = (props) => {
   const { data } = React.useContext(Store);
@@ -10,8 +43,10 @@ const DetailHeader = (props) => {
   return (
     <Header
       {...props}
-      backComponent={<Back />}
       {...useTitle(data, props)}
+      backComponent={
+        <DetailHeaderBackComponent {...props} />
+      }
     />
   );
 };