+++ packages/q3-admin/src/containers/upload/index.jsx
@@ -1,26 +1,15 @@
 import React from 'react';
-import Upload from 'q3-ui/lib/upload';
+import { FileList } from 'q3-ui-filemanager';
 import useRest from 'q3-ui-rest';
-import List from 'q3-ui/lib/list';
 import CircularProgress from '@material-ui/core/CircularProgress';
-import Box from '@material-ui/core/Box';
-import SidePanelContent from '../../components/SidePanelContent';
 import { Definitions } from '../state';
-import FileName from './FileName';
-import FileManage from './FileManage';
 
-const Files = ({ tagOptions, tagInstructions }) => {
+const Upload = () => {
   const { collectionName, id } = React.useContext(
     Definitions,
   );
 
-  const {
-    uploads = [],
-    post,
-    patch,
-    remove,
-    fetching,
-  } = useRest({
+  const { uploads = [], post, remove, fetching } = useRest({
     runOnInit: true,
     url: `/${collectionName}/${id}/uploads`,
     key: 'uploads',
@@ -29,38 +18,15 @@ const Files = ({ tagOptions, tagInstructions }) => {
     },
   });
 
-  return (
-    <SidePanelContent title="fileManager">
-      <Upload fn={post} />
-      {fetching && <CircularProgress />}
-      {uploads.length ? (
-        <List>
-          {uploads.map((file) => (
-            <Box
-              key={file.url}
-              component="li"
-              mb={0.15}
-              p={1}
-              style={{ backgroundColor: '#FFF' }}
-            >
-              <FileName
-                file={file}
-                tagOptions={tagOptions}
-                update={patch(file.id)}
-                {...file}
-              >
-                <FileManage
-                  view={file.url}
-                  remove={remove(file.id)}
-                />
-              </FileName>
-            </Box>
-          ))}
-        </List>
-      ) : null}
-      {tagInstructions}
-    </SidePanelContent>
+  return !fetching ? (
+    <FileList
+      files={uploads}
+      onDelete={remove}
+      onDrop={post}
+    />
+  ) : (
+    <CircularProgress />
   );
 };
 
-export default Files;
+export default Upload;