+++ packages/q3-components/tests/support/mockRequest.js
@@ -1,0 +1,6 @@
+export const simulateNetworkDelay = (values) =>
+  new Promise((r) => {
+    setTimeout(() => {
+      r(values);
+    }, 2500);
+  });