+++ packages/q3-ui-filemanager/src/FileExtensions/index.js
@@ -1,0 +1,6 @@
+import FileExtensions from './FileExtensions';
+
+export default {
+  getColor: FileExtensions('color'),
+  getIcon: FileExtensions('icon'),
+};