+++ packages/q3-ui-filemanager/src/index.js
@@ -1,0 +1,5 @@
+import Avatar from './Avatar';
+import FileList from './FileList';
+import PhotoUpload from './PhotoUpload';
+
+export { Avatar, FileList, PhotoUpload };