+++ packages/q3-ui-forms/src/builders/Wrapper/Wrapper.jsx
@@ -37,6 +37,8 @@ export const InnerForm = ({
   showPersistenceSnack,
   ...etc
 }) => {
+  const [attachments, setAttachments] = React.useState([]);
+
   const {
     isDisabled,
     checkReadAuthorizationContext,
@@ -115,10 +117,13 @@ export const InnerForm = ({
   const execAllSubmitHandlers = onSubmit(() => {
     return forwardProcessStateValuesIntoOnSubmitHandler(
       values,
+      attachments,
     )
       .then((res) => {
         clearPreviousState();
         if (restart) onReset();
+        // always clear the files
+        setAttachments([]);
         return onSuccess(res);
       })
       .catch((e) => {
@@ -156,6 +161,7 @@ export const InnerForm = ({
             initFieldValue,
             setFieldError,
             setFieldValue,
+            setAttachments,
           }}
         >
           <BuilderState.Provider
@@ -166,6 +172,7 @@ export const InnerForm = ({
               isSubmitting,
               errors,
               values,
+              attachments,
             }}
           >
             {children({