+++ packages/q3-ui-forms/CHANGELOG.md
@@ -3,6 +3,53 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.21.5](https://github.com/3merge/q/compare/v1.21.4...v1.21.5) (2020-08-15)
+
+
+### Bug Fixes
+
+* file folder structure for input fields ([12d78f1](https://github.com/3merge/q/commit/12d78f1685dcfa343a48987855c11e8358aad824))
+
+
+
+
+
+## [1.21.4](https://github.com/3merge/q/compare/v1.21.3...v1.21.4) (2020-08-14)
+
+
+### Bug Fixes
+
+* stringify array in formData handler ([2d263d3](https://github.com/3merge/q/commit/2d263d34506cecefbdc25a1ac912f5a090b6edee))
+
+
+
+
+
+## [1.21.3](https://github.com/3merge/q/compare/v1.21.2...v1.21.3) (2020-08-14)
+
+**Note:** Version bump only for package q3-ui-forms
+
+
+
+
+
+## [1.21.2](https://github.com/3merge/q/compare/v1.21.1...v1.21.2) (2020-08-14)
+
+**Note:** Version bump only for package q3-ui-forms
+
+
+
+
+
+## [1.21.1](https://github.com/3merge/q/compare/v1.21.0...v1.21.1) (2020-08-14)
+
+**Note:** Version bump only for package q3-ui-forms
+
+
+
+
+
+
 # [1.21.0](https://github.com/3merge/q/compare/v1.20.22...v1.21.0) (2020-08-07)